Moves all Filter inside the http package
