+++ preview/app/AppLoader.scala
@@ -1,11 +1,12 @@
 import app.{FrontendApplicationLoader, FrontendComponents, LifecycleComponent}
 import com.softwaremill.macwire._
-import conf.{CachedHealthCheckLifeCycle, StandaloneFilters}
+import conf.CachedHealthCheckLifeCycle
 import contentapi.{CapiHttpClient, ContentApiClient, HttpClient}
 import controllers.{HealthCheck, ResponsiveViewerController, StandaloneControllerComponents}
+import http.StandaloneFilters
 import model.ApplicationIdentity
 import play.api.ApplicationLoader.Context
-import play.api._
+import play.api.BuiltInComponentsFromContext
 import play.api.http.HttpErrorHandler
 import play.api.mvc.EssentialFilter
 import play.api.routing.Router
