+++ standalone/app/http/FilterExemptions.scala
@@ -1,4 +1,4 @@
-package conf
+package http
 
 import com.gu.googleauth.FilterExemption
 
@@ -8,9 +8,10 @@ object FilterExemptions {
   lazy val exemptions: Seq[FilterExemption] = List(
     FilterExemption("/oauth2callback"),
     FilterExemption("/assets"),
+    FilterExemption("/favicon.ico"),
     FilterExemption("/_healthcheck"),
-    FilterExemption("/deploys-notify") // This endpoint is authenticated via api-key (to be accessible to riffraff hook for instance)
+    FilterExemption("/2015-06-24-manifest.json"),
+    // the healthcheck url
+    FilterExemption("/world/2012/sep/11/barcelona-march-catalan-independence")
   )
 }
-
-
